feat: add support for postgres. move drop method to test
